+++ src/main/dll/dll_0255_snowbike.c
@@ -532,7 +532,7 @@ void SnowBike_hitDetect(int obj)
         state->unk41C = ((GameObject*)obj)->anim.rotY;
         state->unk41E = ((GameObject*)obj)->anim.rotZ;
         ((GameObject*)obj)->anim.rotY = (f32)((GameObject*)obj)->anim.rotY + state->haloPitchDrift;
-        ((GameObject*)obj)->anim.rotZ = (f32)((GameObject*)obj)->anim.rotZ + ((f32)state->unk410 + state->unk598);
+        ((GameObject*)obj)->anim.rotZ = (f32)((GameObject*)obj)->anim.rotZ + (state->unk410 + state->unk598);
     }
     if (state->unk3D9 == 4 || state->unk3D6 != 0)
     {
